Implement Gogo, add new storages, improve logging

+++ app/src/main/java/ru/rofleksey/animewatcher/storage/TitleStorageEntry.kt
@@ -6,7 +6,7 @@ import ru.rofleksey.animewatcher.api.model.TitleInfo
 class TitleStorageEntry(val info: TitleInfo, val provider: String) : Comparable<TitleStorageEntry> {
     val cachedEpisodeList = mutableListOf<EpisodeInfo>()
     val downloadMap = mutableMapOf<Int, String>()
-    val downloadTasks = mutableMapOf<Int, Long>()
+    val downloadTasks = mutableMapOf<Int, Int>()
     var lastEpisodeNumber: Int = -1
     var lastWatchedEpisode: Int = -1
